更新模型设计,并解决因此导致问题

+++ only4-KSP-cap4j-application/src/main/java/com/only4/application/commands/article/UpdateArticleCommentLikeCountCmd.java
@@ -51,7 +51,7 @@ public class UpdateArticleCommentLikeCountCmd {
     public static class Request implements RequestParam<Response> {
         Long articleId;
         Long commentId;
-        Long likeCount;
+        Integer likeCount;
     }
 
     /**
