+++ only4-KSP-cap4j-application/src/main/java/com/only4/application/commands/article/UpdateArticleLikeCountCmd.java
@@ -50,7 +50,7 @@ public class UpdateArticleLikeCountCmd {
     @AllArgsConstructor
     public static class Request implements RequestParam<Response> {
         Long articleId;
-        Long likeCount;
+        Integer likeCount;
     }
 
     /**
